squash! [DONTMERGE] apps: svc: add link test command.
This should be mergeable now.

Make array static and use memset on it to avoid presumable stack
overflows. This resolves the hardfault.

Set optind to 0 before parsing, to work around the fact that static
variables aren't reinitialized between app invocations.

Increase number of trials per test to 500. The low value of 30 was
chosen to work around I2C hangs; we're on SPI now, which appears to be
more reliable.