greybus-utils: add a way to associate a device_id with a manifest
To setup the routes in SVC, we need to have the device id of source and destination
bridges. Actually, only the source is selectable using kconfig and the destination
is hard coded. To suport more than one manifest, we need to associate device id
with a manifest.

To configure the device id, set CONFIG_DST_DEVID with the device id
(except if you are using custom manifest) or set CONFIG_MANIFEST using the following format:
device_id:manifest_file_name, device_id:manifest_file_name
- device_id is an integer (see configs/ara/svc/src/svc.c for complete list of device) and
- manifest_file_name is the file name of manifest (without extension)

Note: This is not working on endo.

Testing Done:
Bug Number: trivial
Review URL:
Reviewed by: