apbridge app: wait until cport are connected
Currently, apbridge application can try to register cports whereas
they are not yet connected. Because there are no errors handling,
we never retry to register them later and we are using as they were
registered.
Test the cport connection status and wait until cports are connected
(use the manifest to know which cports we must block on).

Testing Done:
Bug Number:
Review URL:
Reviewed by: