Prep for r748 by changing default for legacy bdb defconfigs to SPIROM
Fabien has a patch in queue that allows a "universal" bridge firmware
image to be created that can either be loaded from SPIROM or loaded
into SRAM via JTAG.

For r748 to work and not break the build, we first need to change the
system type for the legacy bdb defconfigs from "load and run from
internal SRAM" to "load from SPI and copy to RAM".

This patch makes the necessary changes to allow r748 to be applied to master.
The sequence will be:
- Apply this patch
- Apply r748
- Update documentation to account for a change in the "load and run from JTAG"
  instructions
- Deprecate the legacy bdb defconfigs and update the documentation to point to
  the new defconfigs

Testing Done: Apply this patch, apply r748, verify that "./build_ara_image.sh
    all" succeeds, manually build bdb/apb1, and verify that the image loads
    via JTAG and runs on AP Bridge A on a BDB1B, using Fabien's recommended
    command sequence.
Bug Number: SW-838
Review URL: https://reviews.leaflabs.com/r/870/
Reviewed by: mbolivar,fparent

Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>