Greybus: add a way to know the driver name or operation name
To help to debug, add a field name for the driver and the operations
in order to know what is happening and which driver or operation is
involved in issue.

Only enabled if CONFIG_GREYBUS_DEBUG is defined.

Testing Done: Build with CONFIG_GREYBUS_DEBUG defined and undefined
Bug Number: SW-617
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>