gpbridge: Don't let the firmware hang if a cport is not connected
The firmware test if a cport is connected before to init and register it.
If the cport is not connected, the firmware waits and seems to have crashed
(no nuttx console). Call the blocking function from another thread.

Testing Done: Tested on GPB1 with SVC configured to routes all cports on APB2: doesn't hang
anymore
Bug Number: SW-941
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>