Greybus-utils: Update Makefile to only include once a manifest
If multiple manifest is in use and we are trying to same manifest twice,
then the manifest file is include twice, causing a build error.
Only include once a manifest.

Testing Done: Build SVC with twice the same manifest. SVC is working and routing table is
correct.
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>