USB gadget: create a new file for common and generic USB gadget code
Create a USB gadget library to make easier APBridgeA's maintenance.
Currently, the library only provide methods to manage requests but
most of the code common or generic code will be moved to this library,
leaving only "vendor" code in gadget.

This library provide a generic request manager.
Requests can be preallocated. If the request pool is empty,
the request manager can allocate a new one.
Note that currently the request manager never free the request
but only return it to the request pool.
Newly allocated are pushed in pool with same size requests.
If there is no request with the same size, a new pool is allocated.

Testing Done: tested with loopback_test.
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>