DWC USB OTG: let the DMa continue onv BNA error
If the SG DMA open a descriptor with a status different from
HOST_READY, then the DMA generate a BNA error.
Once the CPU managed the cause of BNA, the default behavior
of DMA is to restart with the first descriptor.
That is breaking the request management that is expecting circular
operations.
Let the DMA restart with the descriptor that cause the BNA.

Testing Done:
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>