DWC USB OTG: Update BNA handleing
Currently, if the SG DMA generate a BNA,
the CPU reset all descriptors.
This is not a correct way to handle error because of the ring descriptors.
Usualy, BNA is generated because the DMA has used all descriptor and
the CPU didn't get time to complete the transfer and push pack the request
to the list (and make it available to DMA).
The way to handle the BNA is to call completion handler for any completed
transfer and then re enable the endpoint to restart operations.

Testing Done:
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>