usb_test: add an application to test APBridgeA USB gadget
Add an application to test the USB gadget without to
have Greybus or anything else.
The firmware must be built with the dummy backend selected.

Currently, the application send (asynchronous) greybus messages
and wait until it receive them:
if we lost a packet, the application will never exit.

It could be used to quickly determine if the USB works or not.

Testing Done: successfully transfer message to APBridgeA
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>