APBridgeA USB gadget: Not always set or get cport id from Grebus header
Currently, the firmware continue to use header to get or set the cport id
with direct mapped cport id.
Add cportid parameter to some methods and only set it or get it from
header if we are using a muxed enpoint.

Testing Done: Greybus is working (direct and muxed endpoint)
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>