DWC USB OTG: Add method to manage circular list of requests
Add methods to add or remove request from circular list.

dwc_otg_pcd_queue_req() is going to be called on any request
submit. This method should only add the request to circular list
if the request is new (not already in list).
That is expected that USB gadget will submit again and again
the same request because there are preallocated for bulk out endpoints,
and the USB gadget re submit the request on completion.

dwc_otg_pcd_dequeue_req() will only be used if USB gadget
cancel the request, ie if the USB gadget is unloading the USB device
(only true for bulk out endpoints).

Testing Done:
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>