tsb: pll: Allow driver selection for APBridges as well as GPBridges
The new audio architecture uses the I2S port on APBridgeA which
in turn uses the PLL controller when it is the clock master.
Therefore, the Kconfig logic must allow the PLL driver to be
selected for APBridges as well as GPBridges.

Testing Done: Played audio using the White Audio Module
Bug Number: Trivial
Review URL:
Reviewed by:

Signed-off-by: Mark Greer <mgreer@animalcreek.com>