tsb: unipro: re-enable mailbox interrupt in irq handler
During the ES3 bringup I noticed that the GPIO demo was not working with a
ES3 module (with APBridge on it). The mailbox interrupt gets disabled after
the initial mailbox poke for some reasons.
This issue is not present on a ES2 module or with APB2 on the DB3. This commit
workaround this issue by always re-enabling the mailbox IRQ in the interrupt
handler.

Testing Done: gpio demo with ES3 module on DB3.1
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2408/
Reviewed by: eligottlieb, mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>