APBrisgeA USB gadget: Add support of offloaded cport
Use the using mapping table t0 handle offloaded cport.
Add a new value for the cport to endpoint table (OFFLOADED_EP).
If a cport is mapped to this "virtual endpoint" instead of real endpoint,
then the traffic go through two callback registered during the creation of the
offloaded cport.

Before to be offloaded, a cport must be mapped to mixed endpoints
(otherwise, it will keep the direct mapped endpoint in use for nothing).
Cport can be returned to normal operation by mapping it to muexd or direct mapped endpoint.

Testing Done: None because I don't have a driver compatible with offloaded cport.
	      I tried to offload loopback but because loopback doesn't support offloaded cport,
	      greybus destroy the connection.
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>