SW-3008: Fix bridges latency calculation
Entry and exit time are currently attached to one cport.
They are updated for every incoming and outcoming request.
This working fine with synchronous request but it doesn't work with
asynchronous request because entry and exit time may be upadted while
the previous request is still there and so distort its latency.

Testing Done:
	Used gb_loopback_test -t transfer -s 32 -i 1000 -m 0 -p -x -c 8
	and the latency are no bogus anymore.

Change-Id: I12ef7c1911f996d30c72d500b360b30bfdd5251a
Signed-off-by: Alexandre Bailon <abailon@baylibre.com>