SW-3523: DWC OTG PCD: Configure log level with kconfig
Currently, the log level of DWC TOG driver is set to ANY.
When logging is enabled, it can break features due to
timing overhead. Let the user configure it or set it
to OFF.
In addition, update the USB application to configure it at runtime.

Testing Done: Tested with level 0: doesn't break Greybus

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>

Change-Id: Id8944121dce01cc82a43ef8948ebe840e2f22ee8