SW-3518: DWC OTG USB: Ensure bna is handled once
Currently, we are setting the variable bna to let know
to PCD code that a BNA happend and so it have to resume the transfer.
But sometime, we may resume the transfer from BNA interrupt.
In such case, ensure we are not trying to resume again the transfer.

Testing Done: Tested with gb_loopback_test

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>

Change-Id: Ib61384941f2b73db003e5f966d53670eef8424a1