SW-0000: APBridgeA App: remove dummy backend
This backend was useful for testing but it doesn't work anymore
due to some changes.
We still need a way to test USB. The plan is to remove this app
and to use the USB to Cport mapping features to do the samething,
but without the needs to rebuild the firmware to enable it.

Testing Done:
Signed-off-by: Alexandre Bailon <abailon@baylibre.com>

Change-Id: I9bb78f979a10eb9737eb6e15c1a3721d222c3f64