SW-1024: USB Gadget cleanup: Add a way to register a callback to a cport
Currently, we can do that by using the functions for offloading
a cport. But this is too limited so extend it to be more flexible.
The goal is to change the behiavor of transfers for a particular
cport such dump traffic, drop data, etc.

Testing Done: Tested with gb_loopback

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>

Change-Id: I1b4eb0b04b5f6044abdec3e3a27fafda81d22e74